Save basics project after tutorial 3: surfaces
Note that this builds on the previous tutorial.
(Prior to this, all the assets had been separated by namespaces.
After this, the decision was made to simply separate tutorials by commit in version control.
There may be some artifacts as a result of that change.)

Also note that, in addition to the tutorial content, a manager has been added that allows the surface to be changed within the scene using the number keys 1-6.