Save basics project after tutorial 4: performance
As part of the basics tutorial series,
this also builds on the previous tutorial.

The Function Manager used to provide manual control over the graph while in-scene required some additional complexity to account for the different transition mechanisms.
The various transition mechanisms can be manually selected via the following keys:

- 'N' (None)
- 'C' (Cycle)
- 'R' (Random)